REFACTOR: Moved filter types into filter directory - Removed unused type
