+++ app/projects/components/ProjectsList.tsx
@@ -7,8 +7,8 @@ import FilterOverlay from "@/components/Filters/FilterPanel/FilterPanel";
 import ClearAllFiltersButton from "@/components/Filters/Page/ClearAllFiltersButton";
 import ToggleFilterButton from "@/components/Filters/Page/ToggleFilterButton";
 import SearchInput from "@/components/Inputs/SearchInput";
-import FilterCategory from "@/types/FilterCategory";
-import FilterOption from "@/types/FilterOption";
+import FilterCategory from "@/types/filters/FilterCategory";
+import FilterOption from "@/types/filters/FilterOption";
 import Project from "@/types/projects";
 import { Skill } from "@/types/skills";
 import Fuse from "fuse.js";
