--- types/FilterParams.ts
@@ -1,12 +1,0 @@
-interface FilterParams {
-  issuer?: string;
-  category?: string;
-  type?: string;
-  technology?: string;
-  language?: string;
-  search?: string;
-  archived?: boolean;
-  skillCategory?: string;
-}
-
-export default FilterParams;
